Fix astar crash caused by virtual tunnel position
This fix wraps an entities position before executing A*. If an entity is
in the virtual position outside the maze inside the tunnel, the position
is replaced by one side of the tunnel. This will cause an error in the
distance calculation by 1 but is an easy fix for what would otherwise be
an error.